+++ src/frontend/components/Felleskomponenter/SkjemaFeiloppsummering/SkjemaFeiloppsummering.tsx
@@ -13,8 +13,8 @@ import { useApp } from '../../../context/AppContext';
 import { IRoute } from '../../../context/RoutesContext';
 import { IBarn } from '../../../typer/person';
 import { SkjemaFeltTyper } from '../../../typer/skjema';
+import { barnetsNavnValue } from '../../../utils/barn';
 import { samletSpørsmålId, samletSpørsmålSpråkTekstId } from '../../../utils/spørsmål';
-import { barnetsNavnValue } from '../../../utils/visning';
 import { OmDegSpørsmålId } from '../../SøknadsSteg/OmDeg/spørsmål';
 import { VelgBarnSpørsmålId } from '../../SøknadsSteg/VelgBarn/spørsmål';
 import SpråkTekst from '../SpråkTekst/SpråkTekst';
