+++ src/frontend/components/SøknadsSteg/OmBarnet/useOmBarnet.tsx
@@ -36,12 +36,12 @@ import {
     DatoMedUkjent,
     IBarnMedISøknad,
 } from '../../../typer/person';
+import { regexNorskEllerUtenlandskPostnummer } from '../../../utils/adresse';
+import { barnetsNavnValue } from '../../../utils/barn';
 import { validerDatoMedUkjentAvgrensetFremITid } from '../../../utils/dato';
 import { trimWhiteSpace } from '../../../utils/hjelpefunksjoner';
 import { formaterInitVerdiForInputMedUkjent, formaterVerdiForCheckbox } from '../../../utils/input';
 import { svarForSpørsmålMedUkjent } from '../../../utils/spørsmål';
-import { regexNorskEllerUtenlandskPostnummer } from '../../../utils/validering';
-import { barnetsNavnValue } from '../../../utils/visning';
 import SpråkTekst from '../../Felleskomponenter/SpråkTekst/SpråkTekst';
 import { ANNEN_FORELDER } from './SammeSomAnnetBarnRadio';
 import { OmBarnetSpørsmålsId } from './spørsmål';
