+++ src/frontend/components/SøknadsSteg/VelgBarn/Barnekort/Barnekort.tsx
@@ -10,7 +10,7 @@ import { Ingress, Normaltekst, Undertittel } from 'nav-frontend-typografi';
 import { useApp } from '../../../../context/AppContext';
 import { device } from '../../../../Theme';
 import { IBarn } from '../../../../typer/person';
-import { hentBostedSpråkId } from '../../../../utils/person';
+import { hentBostedSpråkId } from '../../../../utils/språk';
 import { formaterFnr } from '../../../../utils/visning';
 import SpråkTekst from '../../../Felleskomponenter/SpråkTekst/SpråkTekst';
 import { TilfeldigBarnIkon } from '../../../Felleskomponenter/TilfeldigBarnIkon/TilfeldigBarnIkon';
