+++ src/frontend/utils/barn.ts
@@ -1,9 +1,21 @@
+import { IntlShape } from 'react-intl';
+
 import { ESvar } from '@navikt/familie-form-elements';
 import { Felt, ISkjema } from '@navikt/familie-skjema';
 
+import { OmBarnaDineSpørsmålId } from '../components/SøknadsSteg/OmBarnaDine/spørsmål';
 import { IOmBarnaDineFeltTyper } from '../components/SøknadsSteg/OmBarnaDine/useOmBarnaDine';
-import { barnDataKeySpørsmål, IBarnMedISøknad } from '../typer/person';
+import { OmBarnetSpørsmålsId } from '../components/SøknadsSteg/OmBarnet/spørsmål';
+import {
+    AlternativtSvarForInput,
+    barnDataKeySpørsmål,
+    barnDataKeySpørsmålUtvidet,
+    IBarn,
+    IBarnMedISøknad,
+    IBarnRespons,
+} from '../typer/person';
 import { ISøknad } from '../typer/søknad';
+import { formaterFnr } from './visning';
 
 export const genererSvarForSpørsmålBarn = (barn: IBarnMedISøknad, felt: Felt<string[]>): ESvar =>
     felt.verdi.includes(barn.id) ? ESvar.JA : ESvar.NEI;
@@ -160,4 +172,186 @@ export const genererOppdaterteBarn = (
             },
         };
     });
+};
+
+export const hentAlder = (dato: string): string => {
+    const idag = new Date();
+    const fødselsdato = new Date(dato);
+    let alder = idag.getFullYear() - fødselsdato.getFullYear();
+    const månedDiff = idag.getMonth() - fødselsdato.getMonth();
+    if (månedDiff < 0 || (månedDiff === 0 && idag.getDate() < fødselsdato.getDate())) {
+        alder--;
+    }
+    return alder.toString();
+};
+
+export const erBarnRegistrertFraFør = (søknad: ISøknad, ident: string) => {
+    const barnFraPdl = søknad.søker.barn.find(barn => barn.ident === ident);
+    const barnRegistrertManuelt = søknad.barnRegistrertManuelt.find(barn => barn.ident === ident);
+    return barnFraPdl || barnRegistrertManuelt;
+};
+
+export const genererInitialBarnMedISøknad = (barn: IBarn): IBarnMedISøknad => {
+    return {
+        ...barn,
+        barnErFyltUt: false,
+        utvidet: {
+            [barnDataKeySpørsmålUtvidet.søkerHarBoddMedAndreForelder]: {
+                id: OmBarnetSpørsmålsId.søkerHarBoddMedAndreForelder,
+                svar: null,
+            },
+            [barnDataKeySpørsmålUtvidet.søkerFlyttetFraAndreForelderDato]: {
+                id: OmBarnetSpørsmålsId.søkerFlyttetFraAndreForelderDato,
+                svar: '',
+            },
+        },
+        [barnDataKeySpørsmål.erFosterbarn]: {
+            id: OmBarnaDineSpørsmålId.hvemErFosterbarn,
+            svar: null,
+        },
+        [barnDataKeySpørsmål.erAdoptertFraUtland]: {
+            id: OmBarnaDineSpørsmålId.hvemErAdoptertFraUtland,
+            svar: null,
+        },
+        [barnDataKeySpørsmål.erAsylsøker]: {
+            id: OmBarnaDineSpørsmålId.hvemErSøktAsylFor,
+            svar: null,
+        },
+        [barnDataKeySpørsmål.barnetrygdFraAnnetEøsland]: {
+            id: OmBarnaDineSpørsmålId.hvemBarnetrygdFraAnnetEøsland,
+            svar: null,
+        },
+        [barnDataKeySpørsmål.oppholderSegIInstitusjon]: {
+            id: OmBarnaDineSpørsmålId.hvemOppholderSegIInstitusjon,
+            svar: null,
+        },
+        [barnDataKeySpørsmål.institusjonsnavn]: {
+            id: OmBarnetSpørsmålsId.institusjonsnavn,
+            svar: '',
+        },
+        [barnDataKeySpørsmål.institusjonsadresse]: {
+            id: OmBarnetSpørsmålsId.institusjonsadresse,
+            svar: '',
+        },
+        [barnDataKeySpørsmål.institusjonspostnummer]: {
+            id: OmBarnetSpørsmålsId.institusjonspostnummer,
+            svar: '',
+        },
+        [barnDataKeySpørsmål.institusjonOppholdStartdato]: {
+            id: OmBarnetSpørsmålsId.institusjonOppholdStartdato,
+            svar: '',
+        },
+        [barnDataKeySpørsmål.institusjonOppholdSluttdato]: {
+            id: OmBarnetSpørsmålsId.institusjonOppholdSluttdato,
+            svar: '',
+        },
+        [barnDataKeySpørsmål.oppholderSegIUtland]: {
+            id: OmBarnaDineSpørsmålId.hvemOppholderSegIUtland,
+            svar: null,
+        },
+        [barnDataKeySpørsmål.oppholdsland]: {
+            id: OmBarnetSpørsmålsId.oppholdsland,
+            svar: '',
+        },
+        [barnDataKeySpørsmål.oppholdslandSluttdato]: {
+            id: OmBarnetSpørsmålsId.oppholdslandSluttdato,
+            svar: '',
+        },
+        [barnDataKeySpørsmål.oppholdslandStartdato]: {
+            id: OmBarnetSpørsmålsId.oppholdslandStartdato,
+            svar: '',
+        },
+        [barnDataKeySpørsmål.boddMindreEnn12MndINorge]: {
+            id: OmBarnaDineSpørsmålId.hvemTolvMndSammenhengendeINorge,
+            svar: null,
+        },
+        [barnDataKeySpørsmål.nårKomBarnTilNorgeDato]: {
+            id: OmBarnetSpørsmålsId.nårKomBarnetTilNorge,
+            svar: '',
+        },
+        [barnDataKeySpørsmål.planleggerÅBoINorge12Mnd]: {
+            id: OmBarnetSpørsmålsId.planleggerÅBoINorge12Mnd,
+            svar: null,
+        },
+        [barnDataKeySpørsmål.barnetrygdFraEøslandHvilketLand]: {
+            id: OmBarnetSpørsmålsId.barnetrygdFraEøslandHvilketLand,
+            svar: '',
+        },
+        [barnDataKeySpørsmål.andreForelderNavn]: {
+            id: OmBarnetSpørsmålsId.andreForelderNavn,
+            svar: '',
+        },
+        [barnDataKeySpørsmål.andreForelderFnr]: {
+            id: OmBarnetSpørsmålsId.andreForelderFnr,
+            svar: '',
+        },
+        [barnDataKeySpørsmål.andreForelderFødselsdato]: {
+            id: OmBarnetSpørsmålsId.andreForelderFødselsdato,
+            svar: '',
+        },
+        [barnDataKeySpørsmål.andreForelderArbeidUtlandet]: {
+            id: OmBarnetSpørsmålsId.andreForelderArbeidUtlandet,
+            svar: null,
+        },
+        [barnDataKeySpørsmål.andreForelderArbeidUtlandetHvilketLand]: {
+            id: OmBarnetSpørsmålsId.andreForelderArbeidUtlandetHvilketLand,
+            svar: '',
+        },
+        [barnDataKeySpørsmål.andreForelderPensjonUtland]: {
+            id: OmBarnetSpørsmålsId.andreForelderPensjonUtland,
+            svar: null,
+        },
+        [barnDataKeySpørsmål.andreForelderPensjonHvilketLand]: {
+            id: OmBarnetSpørsmålsId.andreForelderPensjonHvilketLand,
+            svar: '',
+        },
+        [barnDataKeySpørsmål.borFastMedSøker]: {
+            id: OmBarnetSpørsmålsId.borFastMedSøker,
+            svar: null,
+        },
+        [barnDataKeySpørsmål.skriftligAvtaleOmDeltBosted]: {
+            id: OmBarnetSpørsmålsId.skriftligAvtaleOmDeltBosted,
+            svar: null,
+        },
+        [barnDataKeySpørsmål.søkerForTidsrom]: {
+            id: OmBarnetSpørsmålsId.søkerForTidsrom,
+            svar: null,
+        },
+        [barnDataKeySpørsmål.søkerForTidsromStartdato]: {
+            id: OmBarnetSpørsmålsId.søkerForTidsromStartdato,
+            svar: AlternativtSvarForInput.UKJENT,
+        },
+        [barnDataKeySpørsmål.søkerForTidsromSluttdato]: {
+            id: OmBarnetSpørsmålsId.søkerForTidsromSluttdato,
+            svar: AlternativtSvarForInput.UKJENT,
+        },
+    };
+};
+
+export const hentUid = () => {
+    return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function (c) {
+        const r = (Math.random() * 16) | 0,
+            v = c === 'x' ? r : (r & 0x3) | 0x8;
+        return v.toString(16);
+    });
+};
+
+export const mapBarnResponsTilBarn = (barn: IBarnRespons[]): IBarn[] => {
+    return barn.map(barnRespons => ({
+        id: hentUid(),
+        navn: barnRespons.navn,
+        ident: barnRespons.ident,
+        alder: barnRespons.fødselsdato && hentAlder(barnRespons.fødselsdato),
+        borMedSøker: barnRespons.borMedSøker,
+        adressebeskyttelse: barnRespons.adressebeskyttelse,
+    }));
+};
+
+export const barnetsNavnValue = (barn: IBarn, intl: IntlShape): string => {
+    return barn.adressebeskyttelse
+        ? (intl.formatMessage(
+              { id: 'felles.anonym.barn.fnr' },
+              { fødselsnummer: formaterFnr(barn.ident) }
+          ) as string)
+        : barn.navn.toUpperCase();
 };
