--- src/frontend/utils/person.ts
@@ -1,233 +1,0 @@
-import { OmBarnaDineSpørsmålId } from '../components/SøknadsSteg/OmBarnaDine/spørsmål';
-import { OmBarnetSpørsmålsId } from '../components/SøknadsSteg/OmBarnet/spørsmål';
-import {
-    AlternativtSvarForInput,
-    barnDataKeySpørsmål,
-    barnDataKeySpørsmålUtvidet,
-    ESivilstand,
-    IAdresse,
-    IBarn,
-    IBarnMedISøknad,
-    IBarnRespons,
-    ISamboer,
-    ITidligereSamboer,
-} from '../typer/person';
-import { ISøknad } from '../typer/søknad';
-import { hentUid } from './uuid';
-
-export const hentAlder = (dato: string): string => {
-    const idag = new Date();
-    const fødselsdato = new Date(dato);
-    let alder = idag.getFullYear() - fødselsdato.getFullYear();
-    const månedDiff = idag.getMonth() - fødselsdato.getMonth();
-    if (månedDiff < 0 || (månedDiff === 0 && idag.getDate() < fødselsdato.getDate())) {
-        alder--;
-    }
-    return alder.toString();
-};
-
-export const erBarnRegistrertFraFør = (søknad: ISøknad, ident: string) => {
-    const barnFraPdl = søknad.søker.barn.find(barn => barn.ident === ident);
-    const barnRegistrertManuelt = søknad.barnRegistrertManuelt.find(barn => barn.ident === ident);
-    return barnFraPdl || barnRegistrertManuelt;
-};
-
-const uppercaseFørsteBokstav = text => {
-    if (typeof text !== 'string') return '';
-    return text.charAt(0).toUpperCase() + text.toLowerCase().slice(1);
-};
-
-export const hentAdressefelterSortert = (adresse: IAdresse) => {
-    return [
-        `${adresse.adressenavn ?? ''} ${adresse.husnummer ?? ''}${adresse.husbokstav ?? ''} ${
-            adresse.bruksenhetsnummer ?? ''
-        }`,
-        `${adresse.postnummer ?? ''} ${
-            adresse.poststed ? uppercaseFørsteBokstav(adresse.poststed) : ''
-        }`,
-    ]
-        .map(linje => linje.replace(/\s{2+}/, ' ').trim())
-        .filter(value => value);
-};
-
-export const hentSivilstatus = (statuskode?: ESivilstand) => {
-    switch (statuskode) {
-        case ESivilstand.UGIFT:
-        case ESivilstand.GIFT:
-        case ESivilstand.ENKE_ELLER_ENKEMANN:
-        case ESivilstand.SKILT:
-        case ESivilstand.SEPARERT:
-        case ESivilstand.REGISTRERT_PARTNER:
-        case ESivilstand.SEPARERT_PARTNER:
-        case ESivilstand.SKILT_PARTNER:
-        case ESivilstand.GJENLEVENDE_PARTNER:
-            return `felles.sivilstatus.kode.${statuskode}`;
-
-        default:
-            return `felles.sivilstatus.kode.${ESivilstand.UOPPGITT}`;
-    }
-};
-
-export const genererInitialBarnMedISøknad = (barn: IBarn): IBarnMedISøknad => {
-    return {
-        ...barn,
-        barnErFyltUt: false,
-        utvidet: {
-            [barnDataKeySpørsmålUtvidet.søkerHarBoddMedAndreForelder]: {
-                id: OmBarnetSpørsmålsId.søkerHarBoddMedAndreForelder,
-                svar: null,
-            },
-            [barnDataKeySpørsmålUtvidet.søkerFlyttetFraAndreForelderDato]: {
-                id: OmBarnetSpørsmålsId.søkerFlyttetFraAndreForelderDato,
-                svar: '',
-            },
-        },
-        [barnDataKeySpørsmål.erFosterbarn]: {
-            id: OmBarnaDineSpørsmålId.hvemErFosterbarn,
-            svar: null,
-        },
-        [barnDataKeySpørsmål.erAdoptertFraUtland]: {
-            id: OmBarnaDineSpørsmålId.hvemErAdoptertFraUtland,
-            svar: null,
-        },
-        [barnDataKeySpørsmål.erAsylsøker]: {
-            id: OmBarnaDineSpørsmålId.hvemErSøktAsylFor,
-            svar: null,
-        },
-        [barnDataKeySpørsmål.barnetrygdFraAnnetEøsland]: {
-            id: OmBarnaDineSpørsmålId.hvemBarnetrygdFraAnnetEøsland,
-            svar: null,
-        },
-        [barnDataKeySpørsmål.oppholderSegIInstitusjon]: {
-            id: OmBarnaDineSpørsmålId.hvemOppholderSegIInstitusjon,
-            svar: null,
-        },
-        [barnDataKeySpørsmål.institusjonsnavn]: {
-            id: OmBarnetSpørsmålsId.institusjonsnavn,
-            svar: '',
-        },
-        [barnDataKeySpørsmål.institusjonsadresse]: {
-            id: OmBarnetSpørsmålsId.institusjonsadresse,
-            svar: '',
-        },
-        [barnDataKeySpørsmål.institusjonspostnummer]: {
-            id: OmBarnetSpørsmålsId.institusjonspostnummer,
-            svar: '',
-        },
-        [barnDataKeySpørsmål.institusjonOppholdStartdato]: {
-            id: OmBarnetSpørsmålsId.institusjonOppholdStartdato,
-            svar: '',
-        },
-        [barnDataKeySpørsmål.institusjonOppholdSluttdato]: {
-            id: OmBarnetSpørsmålsId.institusjonOppholdSluttdato,
-            svar: '',
-        },
-        [barnDataKeySpørsmål.oppholderSegIUtland]: {
-            id: OmBarnaDineSpørsmålId.hvemOppholderSegIUtland,
-            svar: null,
-        },
-        [barnDataKeySpørsmål.oppholdsland]: {
-            id: OmBarnetSpørsmålsId.oppholdsland,
-            svar: '',
-        },
-        [barnDataKeySpørsmål.oppholdslandSluttdato]: {
-            id: OmBarnetSpørsmålsId.oppholdslandSluttdato,
-            svar: '',
-        },
-        [barnDataKeySpørsmål.oppholdslandStartdato]: {
-            id: OmBarnetSpørsmålsId.oppholdslandStartdato,
-            svar: '',
-        },
-        [barnDataKeySpørsmål.boddMindreEnn12MndINorge]: {
-            id: OmBarnaDineSpørsmålId.hvemTolvMndSammenhengendeINorge,
-            svar: null,
-        },
-        [barnDataKeySpørsmål.nårKomBarnTilNorgeDato]: {
-            id: OmBarnetSpørsmålsId.nårKomBarnetTilNorge,
-            svar: '',
-        },
-        [barnDataKeySpørsmål.planleggerÅBoINorge12Mnd]: {
-            id: OmBarnetSpørsmålsId.planleggerÅBoINorge12Mnd,
-            svar: null,
-        },
-        [barnDataKeySpørsmål.barnetrygdFraEøslandHvilketLand]: {
-            id: OmBarnetSpørsmålsId.barnetrygdFraEøslandHvilketLand,
-            svar: '',
-        },
-        [barnDataKeySpørsmål.andreForelderNavn]: {
-            id: OmBarnetSpørsmålsId.andreForelderNavn,
-            svar: '',
-        },
-        [barnDataKeySpørsmål.andreForelderFnr]: {
-            id: OmBarnetSpørsmålsId.andreForelderFnr,
-            svar: '',
-        },
-        [barnDataKeySpørsmål.andreForelderFødselsdato]: {
-            id: OmBarnetSpørsmålsId.andreForelderFødselsdato,
-            svar: '',
-        },
-        [barnDataKeySpørsmål.andreForelderArbeidUtlandet]: {
-            id: OmBarnetSpørsmålsId.andreForelderArbeidUtlandet,
-            svar: null,
-        },
-        [barnDataKeySpørsmål.andreForelderArbeidUtlandetHvilketLand]: {
-            id: OmBarnetSpørsmålsId.andreForelderArbeidUtlandetHvilketLand,
-            svar: '',
-        },
-        [barnDataKeySpørsmål.andreForelderPensjonUtland]: {
-            id: OmBarnetSpørsmålsId.andreForelderPensjonUtland,
-            svar: null,
-        },
-        [barnDataKeySpørsmål.andreForelderPensjonHvilketLand]: {
-            id: OmBarnetSpørsmålsId.andreForelderPensjonHvilketLand,
-            svar: '',
-        },
-        [barnDataKeySpørsmål.borFastMedSøker]: {
-            id: OmBarnetSpørsmålsId.borFastMedSøker,
-            svar: null,
-        },
-        [barnDataKeySpørsmål.skriftligAvtaleOmDeltBosted]: {
-            id: OmBarnetSpørsmålsId.skriftligAvtaleOmDeltBosted,
-            svar: null,
-        },
-        [barnDataKeySpørsmål.søkerForTidsrom]: {
-            id: OmBarnetSpørsmålsId.søkerForTidsrom,
-            svar: null,
-        },
-        [barnDataKeySpørsmål.søkerForTidsromStartdato]: {
-            id: OmBarnetSpørsmålsId.søkerForTidsromStartdato,
-            svar: AlternativtSvarForInput.UKJENT,
-        },
-        [barnDataKeySpørsmål.søkerForTidsromSluttdato]: {
-            id: OmBarnetSpørsmålsId.søkerForTidsromSluttdato,
-            svar: AlternativtSvarForInput.UKJENT,
-        },
-    };
-};
-
-export const mapBarnResponsTilBarn = (barn: IBarnRespons[]): IBarn[] => {
-    return barn.map(barnRespons => ({
-        id: hentUid(),
-        navn: barnRespons.navn,
-        ident: barnRespons.ident,
-        alder: barnRespons.fødselsdato && hentAlder(barnRespons.fødselsdato),
-        borMedSøker: barnRespons.borMedSøker,
-        adressebeskyttelse: barnRespons.adressebeskyttelse,
-    }));
-};
-
-export const hentBostedSpråkId = (barn: IBarn) => {
-    if (barn.adressebeskyttelse) {
-        return 'hvilkebarn.barn.bosted.adressesperre';
-    } else if (barn.borMedSøker) {
-        return 'hvilkebarn.barn.bosted.din-adresse';
-    } else {
-        return 'hvilkebarn.barn.bosted.ikke-din-adresse';
-    }
-};
-
-export const erTidligereSamboer = (
-    samboer: ISamboer | ITidligereSamboer
-): samboer is ITidligereSamboer => {
-    return 'samboerTilDato' in samboer;
-};
