--- src/frontend/utils/validering.ts
@@ -1,2 +1,0 @@
-export const regexNorskEllerUtenlandskPostnummer = (verdi: string) =>
-    verdi.match(/^[A-Za-z0-9-æøåÆØÅ-]{3,10}$/);
