+++ src/frontend/utils/visning.ts
@@ -1,0 +1,17 @@
+import { AlternativtSvarForInput, DatoMedUkjent } from '../typer/person';
+import { formaterDato } from './dato';
+
+export const formaterFnr = (fødselsnummer: string) => {
+    return fødselsnummer.substring(0, 6) + ' ' + fødselsnummer.substring(6, 11);
+};
+
+export const formaterDatoMedUkjent = (datoMedUkjent: DatoMedUkjent, tekstForUkjent): string => {
+    return datoMedUkjent === AlternativtSvarForInput.UKJENT
+        ? tekstForUkjent
+        : formaterDato(datoMedUkjent);
+};
+
+export const uppercaseFørsteBokstav = text => {
+    if (typeof text !== 'string') return '';
+    return text.charAt(0).toUpperCase() + text.toLowerCase().slice(1);
+};
